ajout de boucle pour remplir tout le template

+++ fill_template.py
@@ -2,9 +2,9 @@ import os
 import openai
 from dotenv import load_dotenv, find_dotenv
 import pandas as pd
-import numpy as np
 import subprocess
 import gradio as gr
+import time
 import vectorstore_lib
 import call_to_llm
 
@@ -12,22 +12,18 @@ from langchain.document_loaders import PyMuPDFLoader
 from langchain.chat_models import ChatOpenAI
 from langchain.embeddings.openai import OpenAIEmbeddings
 from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain.vectorstores import Chroma
 from langchain.prompts import PromptTemplate
-from langchain.chains import LLMChain
 
 _ = load_dotenv(find_dotenv()) # read local .env file
 openai.api_key  = os.environ['OPENAI_API_KEY']
 
-cv_path = "data/cv/"
+template_path = "data_template_concise.csv"
+cv_path = "data/"
 complete_paths = [cv_path+file for file in os.listdir(cv_path) if 'pdf' in file]
 
-## Loading the info template
-template_df = pd.read_csv("dataTemplateConcise.csv")
-fields = list(template_df.iloc[:,1:]) # Avoiding the index field
 ## Choice of the prompt template to feed the LLM
 
-# Prompt asking to directly fill a template
+# Prompt asking to fill directly a template
 prompt_global = """ The texts provided to you are the resumes of the candidates.
 Your task is to answer the user question in a structured way as in the following format.
 
@@ -71,7 +67,7 @@ Curriculum Vitae : ```{context}```
 """
 
 # Different prompt for each field
-prompts_df = pd.read_csv("promptTemplatesConcise.csv")
+#prompts_df = pd.read_csv("promptTemplatesConcise.csv")
 
 ########## LLM Chain parameters ##########
 
@@ -87,44 +83,63 @@ nb_chunks = 4
 llm_name = 'gpt-3.5-turbo'
 llm = ChatOpenAI(model_name=llm_name, temperature=0)
 
-def fill_template(file, save=False):
+def fill_one_row(template_path, file, save=False, verbose=False):
+
+    template_df = pd.read_csv(template_path)
+    fields = list(template_df) # Avoiding the index field
     
-    loader = PyMuPDFLoader(file)
-    cv = loader.load()
-    data = [file] # First field is the filename
+    if file in template_df["Filename"].unique():
+        print("This CV has already been parsed")
+        return template_df.loc[template_df["Filename"] == "data/CV_robert_vesoul.pdf"].values.tolist()[0]
+    
+    else:
 
-    ## Vectorstore creation
+        loader = PyMuPDFLoader(file)
+        cv = loader.load()
+        data = [file] # First field is the filename
 
-    subprocess.run('rm -rf ./chroma_single/', shell=True)
-    chunks = vectorstore_lib.splitting_of_docs(cv, splitter)
-    vectordb = vectorstore_lib.new_vectordb(chunks, 
-                            embedding, 
-                            persist_directory
-    )    
+        ## Vectorstore creation
 
-    for field in fields[1:]:    # Omitting the first field which corresponds to the filename
+        subprocess.run('rm -rf ./chroma_single/', shell=True)
+        chunks = vectorstore_lib.splitting_of_docs(cv, splitter)
+        vectordb = vectorstore_lib.new_vectordb(chunks, 
+                                embedding, 
+                                persist_directory
+        )    
 
-        ## Prepare chain with prompt template
+        for field in fields[1:]:    # Omitting the first field (=filename)
 
-        # prompt_template = prompts_df[field][0]
-        # prompt = PromptTemplate(template=prompt_template, input_variables=["context"])
-        prompt = PromptTemplate(template=prompt_generic, input_variables=["context","field"])
-        chain = call_to_llm.create_chain(llm, prompt)
+            ## Prepare chain with prompt template
 
-        ## Retrieving and calling the LLM
-        sources = vectordb.similarity_search(field, k=nb_chunks)
-        context = call_to_llm.create_context_from_retrieved_chunks(sources)
-        answer = chain.predict(context=context, field=field)
-        data.append(answer)
-        #data.append(singleCV_Extraction.gradioQA(cv, "What is the " + field + " of the candidate ?", with_source=False, print_prompt=False)[0])
-    df = pd.DataFrame(columns=fields, data=[data])
-    if save:
-        template_df.append(df)
-#    return df
-    return list(df.iloc[0])
+            # prompt_template = prompts_df[field][0]
+            # prompt = PromptTemplate(template=prompt_template, input_variables=["context"])
+            prompt = PromptTemplate(template=prompt_generic, input_variables=["context","field"])
+            chain = call_to_llm.create_chain(llm, prompt)
+
+            ## Retrieving and calling the LLM
+            sources = vectordb.similarity_search(field, k=nb_chunks)
+            context = call_to_llm.create_context_from_retrieved_chunks(sources)
+            answer = chain.predict(context=context, field=field)
+            data.append(answer)
+            if verbose:
+                print(data)
+        df = pd.DataFrame(columns=fields, data=[data])
+        if save:
+            result_df = pd.concat([template_df, df], ignore_index=True)
+            result_df.to_csv(template_path, index=False)
+        return list(df.iloc[0])
+
+def fill_whole_template(template_path, complete_paths):
+    t0 = time.time()
+    for file in complete_paths:
+        try:
+            fill_one_row(template_path, file, save=True, verbose=True)
+        except:
+            print("Error filling the template for " + file)
+    return time.time()-t0
 
 demo = gr.Interface(
-    fn = lambda file: fill_template(file),
+    fn = lambda file: fill_one_row(template_path, file),
     inputs =
     [
         gr.Dropdown(
@@ -133,15 +148,14 @@ demo = gr.Interface(
     ],
     outputs=
     [
-        #gr.Dataframe(label="Candidate Information")
-        gr.Textbox(label = field) for field in fields
+        gr.Textbox(label = field) for field in list(pd.read_csv(template_path))
     ],
     description = "Analyzes a file according to the given template"
 )
 
-if __name__ == "__main__":
-    demo.launch(inbrowser=True)
+#demo.launch(inbrowser=True)
 
-
-# file = complete_paths[0]
-# print(fill_template(file, save=True))
+# print(complete_paths)
+# file = complete_paths[5]
+# print(fill_one_row(template_path, file, save=True))
+print(fill_whole_template(template_path, complete_paths))
